Fix bootstrapper tests, add mediator tests and update code coverage
