+++ MessagingService.BusinessLogic.Tests/Mediator/MediatorTests.cs
@@ -1,27 +1,35 @@
-﻿using System;
+﻿using MediatR;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+using Moq;
+using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
+using Xunit;
 
-namespace MessagingService.Tests.General
+namespace MessagingService.BusinessLogic.Tests.Mediator
 {
-    using System.Diagnostics;
-    using System.Linq;
+    using BusinessLogic.Services;
     using Lamar;
-    using Microsoft.AspNetCore.Hosting;
-    using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Hosting;
-    using Moq;
-    using Xunit;
+    using Testing;
 
-    [Collection("TestCollection")]
-    public class BootstrapperTests
+    public class MediatorTests
     {
-        #region Methods
+        private List<IBaseRequest> Requests = new List<IBaseRequest>();
+
+        public MediatorTests() {
+            this.Requests.Add(TestData.ResendEmailRequest);
+            this.Requests.Add(TestData.SendEmailRequestNoAttachments);
+            this.Requests.Add(TestData.SendSMSRequest);
+        }
 
         [Fact]
-        public void VerifyBootstrapperIsValid()
-        {
+        public async Task Mediator_Send_RequestHandled() {
             Mock<IWebHostEnvironment> hostingEnvironment = new Mock<IWebHostEnvironment>();
             hostingEnvironment.Setup(he => he.EnvironmentName).Returns("Development");
             hostingEnvironment.Setup(he => he.ContentRootPath).Returns("/home");
@@ -34,16 +42,29 @@ namespace MessagingService.Tests.General
             this.AddTestRegistrations(services, hostingEnvironment.Object);
             s.ConfigureContainer(services);
             Startup.Container.AssertConfigurationIsValid(AssertMode.Full);
+
+            List<String> errors = new List<String>();
+            IMediator mediator = Startup.Container.GetService<IMediator>();
+            foreach (IBaseRequest baseRequest in this.Requests) {
+                try {
+                    await mediator.Send(baseRequest);
+                }
+                catch(Exception ex) {
+                    errors.Add(ex.Message);
+                }
+            }
+
+            if (errors.Any() == true) {
+                String errorMessage = String.Join(Environment.NewLine, errors);
+                throw new Exception(errorMessage);
+            }
         }
 
-        private IConfigurationRoot SetupMemoryConfiguration()
-        {
+        private IConfigurationRoot SetupMemoryConfiguration() {
             Dictionary<String, String> configuration = new Dictionary<String, String>();
 
             IConfigurationBuilder builder = new ConfigurationBuilder();
 
-            configuration.Add("AppSettings:EmailProxy", "UnitTest");
-            configuration.Add("AppSettings:SMSProxy", "UnitTest");
             configuration.Add("ConnectionStrings:HealthCheck", "HeathCheckConnString");
             configuration.Add("SecurityConfiguration:Authority", "https://127.0.0.1");
             configuration.Add("EventStoreSettings:ConnectionString", "https://127.0.0.1:2113");
@@ -53,29 +74,27 @@ namespace MessagingService.Tests.General
             configuration.Add("AppSettings:UseConnectionStringConfig", "false");
             configuration.Add("AppSettings:SecurityService", "http://127.0.0.1");
             configuration.Add("AppSettings:MessagingServiceApi", "http://127.0.0.1");
+            configuration.Add("AppSettings:TransactionProcessorApi", "http://127.0.0.1");
+            configuration.Add("AppSettings:DatabaseEngine", "SqlServer");
+            configuration.Add("AppSettings:EmailProxy", "UnitTest");
+            configuration.Add("AppSettings:SMSProxy", "UnitTest");
 
             builder.AddInMemoryCollection(configuration);
 
             return builder.Build();
         }
 
-        /// <summary>
-        /// Adds the test registrations.
-        /// </summary>
-        /// <param name="services">The services.</param>
-        /// <param name="hostingEnvironment">The hosting environment.</param>
-        private void AddTestRegistrations(IServiceCollection services,
-                                          IWebHostEnvironment hostingEnvironment)
-        {
+        private void AddTestRegistrations(ServiceRegistry services,
+                                          IWebHostEnvironment hostingEnvironment) {
             services.AddLogging();
             DiagnosticListener diagnosticSource = new DiagnosticListener(hostingEnvironment.ApplicationName);
             services.AddSingleton<DiagnosticSource>(diagnosticSource);
-            services.AddSingleton(diagnosticSource);
+            services.AddSingleton<DiagnosticListener>(diagnosticSource);
             services.AddSingleton<IWebHostEnvironment>(hostingEnvironment);
             services.AddSingleton<IHostEnvironment>(hostingEnvironment);
             services.AddSingleton<IConfiguration>(Startup.Configuration);
-        }
 
-        #endregion
+            services.OverrideServices(s => { s.AddSingleton<IMessagingDomainService, DummyMessagingDomainService>(); });
+        }
     }
 }
