+++ MessagingService.BusinessLogic.Tests/MessagingService.BusinessLogic.Tests.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Lamar" Version="10.0.1" />
     <PackageReference Include="Moq" Version="4.18.2" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
     <PackageReference Include="Shouldly" Version="4.1.0" />
@@ -24,6 +25,7 @@
   <ItemGroup>
     <ProjectReference Include="..\MessagingService.BusinessLogic\MessagingService.BusinessLogic.csproj" />
     <ProjectReference Include="..\MessagingService.Testing\MessagingService.Testing.csproj" />
+    <ProjectReference Include="..\MessagingService\MessagingService.csproj" />
   </ItemGroup>
 
 </Project>
