+++ MessagingService.sln
@@ -35,15 +35,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagingService.SMSMessage
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagingService.SMSAggregate.Tests", "MessagingService.SMSAggregate.Tests\MessagingService.SMSAggregate.Tests.csproj", "{883DEFD0-7DDB-4F79-9996-2279E71CAF5B}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F7B72EF7-16B2-489A-8233-ED237A67DA51}"
-	ProjectSection(SolutionItems) = preProject
-		.github\workflows\buildwindowsimage.yml = .github\workflows\buildwindowsimage.yml
-		.github\workflows\createrelease.yml = .github\workflows\createrelease.yml
-		.github\workflows\nightlybuild.yml = .github\workflows\nightlybuild.yml
-		.github\workflows\pullrequest.yml = .github\workflows\pullrequest.yml
-		.github\workflows\pushtomaster.yml = .github\workflows\pushtomaster.yml
-	EndProjectSection
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
