+++ MessagingService/Bootstrapper/DomainServiceRegistry.cs
@@ -3,7 +3,9 @@
 using BusinessLogic.Services;
 using Lamar;
 using Microsoft.Extensions.DependencyInjection;
+using System.Diagnostics.CodeAnalysis;
 
+[ExcludeFromCodeCoverage]
 public class DomainServiceRegistry : ServiceRegistry
 {
     public DomainServiceRegistry()
