+++ MessagingService/Bootstrapper/MediatorRegistry.cs
@@ -1,6 +1,7 @@
 ﻿namespace MessagingService.Bootstrapper;
 
 using System;
+using System.Diagnostics.CodeAnalysis;
 using BusinessLogic.RequestHandlers;
 using BusinessLogic.Requests;
 using Lamar;
@@ -8,6 +9,7 @@ using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Shared.General;
 
+[ExcludeFromCodeCoverage]
 public class MediatorRegistry : ServiceRegistry
 {
     public MediatorRegistry()
