+++ MessagingService/MessagingService.csproj
@@ -10,8 +10,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="Lamar" Version="8.1.0" />
-	  <PackageReference Include="Lamar.Microsoft.DependencyInjection" Version="8.1.0" />
+	  <PackageReference Include="Lamar" Version="10.0.1" />
+	  <PackageReference Include="Lamar.Microsoft.DependencyInjection" Version="10.0.1" />
 	  <PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="6.0.5" />
     <PackageReference Include="AspNetCore.HealthChecks.Uris" Version="6.0.3" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.0" />
